Drumov Maxim Block4

+++ DrumovMaxim/Semester1/Block4/Task1/Task1/bigint.h
@@ -7,12 +7,12 @@
 typedef struct bigInt
 {
 	size_t size;
-	uint64_t* digits;
+	uint32_t* digits;
 }bigInt;
 
-bigInt* newBigInt(size_t size, uint64_t value);
+bigInt* newBigInt(size_t size, uint32_t value);
 void freeBigInt(bigInt* number);
 void printfBigIntHex(bigInt* number);
 bigInt* multiply(bigInt* left, bigInt* right);
-bigInt* power(bigInt* number, uint64_t pow);
+bigInt* power(bigInt* number, int32_t pow);
